refactor(#2869): extract the two fat-jar oracles into gradle/fatjar-oracles.gradle.kts
Owner ruling on the 884-line flag. Split by RESPONSIBILITY: verification moves
out, build configuration stays.

  trino-connector/build.gradle.kts            884 -> 464 lines
  trino-connector/gradle/fatjar-oracles.gradle.kts  (new)  477 lines

Stays in build.gradle.kts, deliberately: the `shadow { }` block, `shadowJar`'s
transformers + duplicatesStrategy bypass, the no-relocation rationale, and
`installPluginFat` — those are configuration, not verification.
`verifyPublishedPomNettyPin` (#2300) also stays: pre-existing, and moving it
would enlarge the diff for nothing.

`apply(from = ...)` compiles the applied script SEPARATELY, so two sharp edges
are handled explicitly rather than discovered later:

  * No type-safe accessors across the boundary. `configurations.runtimeClasspath`
    -> `configurations.named("runtimeClasspath")`, `dependsOn(tasks.shadowJar)` ->
    `dependsOn("shadowJar")`. Task NAMES and the `check` wiring are unaffected —
    both tasks register on the same Project, so `./gradlew shadowJar verifyFatJar
    verifyShadowNotPublished` (Unit D's release workflow, Unit C's CI lane)
    resolves exactly as before and no task was renamed.
  * build.gradle.kts's `nettyVersion` / `nettyTcnativeVersion` / `arrowVersion`
    locals are NOT visible to it. They are passed through `extra` from their single
    declaration site rather than re-typed as literals: a second copy of a pinned
    version is precisely the drift verifyPublishedPomNettyPin exists to catch, and
    would let the oracle certify a pin the build no longer uses. A missing extra
    throws; it cannot silently default.

BEHAVIOUR UNCHANGED — location only. Also drops the now-unused java.io/java.util
imports from build.gradle.kts (they existed solely for the oracles).

Both files land in ONE commit: `apply(from = ...)` referencing an untracked file
would break Gradle CONFIGURATION on a fresh checkout, taking down every step in
trino-connector-ci.yml including the pre-existing `test` step, and both oracles
would exist nowhere in the repo. Verified with `git ls-tree HEAD`, not
`git status` — an untracked new file is invisible to a `git add <known-paths>`
habit.
